Allow confusion-making.
git-svn-id: https://lisatools.googlecode.com/svn@644 c0d0b04f-df38-0410-a060-df9e05943d0e

+++ MLDCpipelines2/bin/makesource-Galaxy3.py
@@ -34,8 +34,12 @@ parser.add_option("-s", "--seed",
                   help="Galaxy seed (int) [required]")
 
 parser.add_option("-v", "--verification",
-                  action="store_true", dest="verification", default=False,
-                  help="include verification binaries [defaults to true, currently not active for Challenge 3]")
+                  action="store_true", dest="verification", default=True,
+                  help="include verification binaries [defaults to true]")
+
+parser.add_option("-c", "--confusion",
+                  action="store_true", dest="confusion", default=False,
+                  help="run Galaxy3 code to produce partially regressed confusion foreground [defaults to False]")
 
 (options, args) = parser.parse_args()
 
@@ -44,6 +48,10 @@ if options.seed == None:
 
 seed = options.seed
 verification = options.verification
+confusion = options.confusion
+
+if verification and confusion:
+    parser.error("The --verification and --confusion options are incompatible.")
     
 if len(args) != 1:
     parser.error("I need a location for the output Galaxy file!")
@@ -65,7 +73,7 @@ here = os.getcwd()
 workdir = os.path.abspath(tempfile.mkdtemp(dir='.'))
 os.chdir(workdir)
 
-execfiles = ['Fast_Response3','Fast_XML_LS3','Fast_XML_SL3','Galaxy_key3','Galaxy_Maker3']
+execfiles = ['Fast_Response3','Fast_XML_LS3','Fast_XML_SL3','Galaxy_key3','Galaxy_Maker3','Confusion_Maker3']
 
 for f in execfiles:
     run('ln -s %s/%s ./%s' % (galaxydir,f,f),quiet=True)
@@ -76,6 +84,7 @@ run('mkdir XML',quiet=True)
 
 run('ln -s %s/Data/AMCVn_GWR_MLDC.dat ./Data/AMCVn_GWR_MLDC.dat' % galaxydir,quiet=True)
 run('ln -s %s/Data/dwd_GWR_MLDC.dat ./Data/dwd_GWR_MLDC.dat' % galaxydir,quiet=True)
+run('ln -s %s/Data/Verification.dat ./Data/Verification.dat' % galaxydir,quiet=True)
 
 # make Galaxy!
 
@@ -87,6 +96,8 @@ os.chdir(workdir)
 
 if verification:
     run('./Galaxy_Maker3 %s 1' % seed)
+elif confusion:
+    run('./Confusion_Maker3 %s' % seed)
 else:
     run('./Galaxy_Maker3 %s 0' % seed)
 
@@ -98,7 +109,9 @@ run("sed 's/Data\///g' < XML/TheGalaxy_key.xml > %s" % galaxyfile)
 
 destdir = os.path.dirname(galaxyfile) + '/.'
 run("mv Data/Galaxy_%s.dat %s" % (seed,destdir))
-run("mv Data/Galaxy_Bright_%s.dat %s" % (seed,destdir))
+# if we do confusion, this file may not be here
+if os.path.isfile('Data/Galaxy_Bright_%s.dat' % seed):
+    run("mv Data/Galaxy_Bright_%s.dat %s" % (seed,destdir))
 
 run('rm -f XML/noise-only.xml',quiet=True)
 run('rm -f XML/noise-only.bin',quiet=True)
